+++ test/support/my_mailer.rb
@@ -1,0 +1,2 @@
+class MyMailer < Devise::Mailer
+end