+++ test/test_helper.rb
@@ -12,21 +12,8 @@ require "resque"
 require "sidekiq"
 require "delayed_job_active_record"
 
+require "support/rails_app"
 require "support/helpers"
+require "support/my_mailer"
 
-# Silent schema load output
-ActiveRecord::Migration.verbose = false
-
-module Devise
-  module Async
-    class RailsApp < ::Rails::Application
-      config.root = File.dirname(__FILE__) + "/support/rails_app"
-      config.active_support.deprecation = :log
-      config.action_mailer.default_url_options = { :host => "localhost:3000" }
-      config.action_mailer.delivery_method = :test
-    end
-  end
-end
-
-Devise::Async::RailsApp.initialize!
 load File.dirname(__FILE__) + "/support/rails_app/db/schema.rb"