+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+* New
+	* Added `Devise::Async.mailer` option to proxy to custom mailers
+	* Added `Devise::Async.setup` to allow configuring with blocks
+
 ## 0.0.2
 
 * Enhancements
@@ -6,7 +12,6 @@
 * Deprecations
 	* DeviseAsync::Proxy is now Devise::Async::Proxy
 	* DeviseAsync.backend is now Devise::Async.backend
-
 
 ## 0.0.1